Product Frontend

+++ controllers/category.js
@@ -1,4 +1,5 @@
 const Category = require("../models/categoryModel")
+const Products = require("../models/product")
 
 const categoryCtrl = {
     getCategories,
@@ -36,6 +37,8 @@ async function createCategory(req, res) {
 
 async function deleteCategory(req, res) {
     try {
+        const products = await Products.findOne({ category: req.params.id })
+        if (products) return res.status(400).json({ msg: 'Please delete all products with a relationship.' })
         await Category.findByIdAndDelete(req.params.id)
         res.json({ msg: "Deleted a category" })
